+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -212,6 +212,10 @@ New checks
 
 - New :doc:`daedalean-switch-statement
   <clang-tidy/checks/daedalean-switch-statement>` check.
+
+- New :doc:`daedalean-ternary-operator-must-not-be-used
+  <clang-tidy/checks/daedalean-ternary-operator-must-not-be-used>` check.
+
 - New :doc:`daedalean-lambda-return-type
   <clang-tidy/checks/daedalean-lambda-return-type>` check.